feat: Added default dates and times functionality
This commit adds a button to set the input fields for start and end dates and times to their default values. The `setDefaultDatesAndTimes()` function is called when the button is clicked and it sets the input values to the default start and end dates and times. Additionally, the `DOMContentLoaded` event listener is used to check if the filters are set in the URL parameters, and if not, it sets the input fields to their default values.